Crate ic_nns_gtc

Modules

init

pb

test_constants

Constants

LOG_PREFIX

SECONDS_UNTIL_CLAIM_NEURONS_CAN_BE_CALLED
The amount of time after the genesis of the IC that GTC neurons cannot be claimed.

SECONDS_UNTIL_FORWARD_ALL_UNCLAIMED_ACCOUNTS_CAN_BE_CALLED
The amount of time after the genesis of the IC that any user can call forward_all_unclaimed_accounts. This allows the reclaiming of GTC neurons that have not been claimed, so that these neurons don't exist in an unclaimed state indefinitely.

Functions

der_encode
DER-encode the given PublicKey